Get rid of some warnings

+++ usvdata/Path.cpp
@@ -285,7 +285,7 @@ std::vector<Vector2> Path::getPointsPath(const double angle_increment)const{
             double r = std::abs(1 / s._curve);
             auto dangle = std::abs(s._length * s._curve);
             double sign = s._curve>0?1:-1;
-            size_t n=dangle/angle_increment;
+            size_t n = static_cast<size_t>(dangle/angle_increment);
             for(size_t i=0;i<n;++i){
                 auto x_ = std::sin(angle_increment*i);
                 auto y_ = sign * (1 - std::cos(angle_increment*i));
